minor fixes after merge with comment-skuli

+++ src/main/java/is/hi/hbv501g/team20/taeknilaesi/model/User.java
@@ -38,11 +38,11 @@ public class User
         this.password = password;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
     public String getEmail() {
